The Update README.md
README.md has been updated with a thorough explanation of the code's operation.
added a thorough description of the goal, prerequisites for input, computations, and results of the program. Information regarding the use of the Scanner class for user input is also supplied. This patch improves the code's readability and comprehension for upcoming users and developers.